+++ CHANGELOG.md
@@ -1,3 +1,34 @@
+v.1.0.4
+------
+- [x] Add module `churn_predictor_rest_api.py` in `src/model_deployment`
+- [x] Add `Dockerfile` in `docker`
+
+v.1.0.3
+------
+- [x] Add module `model_training_utils.py` in `src/model_training`
+- [x] Add module `model_training.py` in `src/model_training`
+- [x] Add PyTest `test_model_training.py` in `tests`
+- [x] Add Notebook `telco_customer_churn_model_training.ipynb` in `notebooks`
+
+v.1.0.2
+------
+- [x] Add module `data_preparation_utils.py` in `src/data_preparation`
+- [x] Add `conftest.py` in `tests`
+- [x] Add PyTest `test_data_preparation.py` in `tests`
+- [x] Add module `data_preparation` in `src/data_preparation`
+- [x] Add Notebook `telco_customer_churn_model_training.ipynb` in `notebooks`
+
+v.1.0.1
+------
+- [x] Add data `telco_customer_churn_data.csv` in `data/raw`
+- [x] Add function `get_logger` in `src/logging_module/logging_module.py`
+- [x] Add `log_configuration.yaml` in. `src/logging_module`
+- [x] Add PyTest `test_logging_module` in `tests`
+- [x] Add function `read_configuration` in `src/general_utils/general_utils.py`
+- [x] Add PyTest `test_general_utils` in `tests`
+- [x] Add `config.yaml` in `configuration`
+- [x] Add Jupyter Notebook `telco_customer_churn_eda.ipynb` in `notebooks`
+
 v.0.1.0
 ------
 - [x] Setup Poetry